fixes from code review
