+++ shared/ViewComponents/CourseDetailsViewComponent.cs
@@ -36,19 +36,4 @@ namespace GovUk.Education.SearchAndCompare.UI.Shared.ViewComponents
             return View(viewModel);
         }
     }
-
-    public static class CourseDetailsSections
-    {
-        public const string AboutTheCourse = "about this training programme";
-        public const string InterviewProcess = "interview process";
-        public const string EntryRequirementsQualifications = "entry requirements";
-        public const string EntryRequirementsPersonalQualities = "entry requirements personal qualities";
-        public const string EntryRequirementsOther = "entry requirements other";
-        public const string ListOfSchools = "placement schools";
-        public const string AboutSchools = "about school placements";
-        public const string AboutFees = "about fees";
-        public const string AboutTheProvider = "about this training provider";
-        public const string AboutTheAccreditingProvider = "about this training provider accrediting";        
-        public const string TrainWithDisabilities = "training with disabilities";        
-    }
 }
